requirements.txt
django>=1.5
wheel==0.24
django-ordered-model==0.4